% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{as_kreuzberg_result}
\alias{as_kreuzberg_result}
\title{Convert a list to a kreuzberg_result S3 object}
\usage{
as_kreuzberg_result(x)
}
\arguments{
\item{x}{A named list from native extraction.}
}
\value{
Object with class \code{kreuzberg_result}.
}
\description{
Convert a list to a kreuzberg_result S3 object
}
\keyword{internal}
