% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{clear_document_extractors}
\alias{clear_document_extractors}
\title{Clear all registered document extractors}
\usage{
clear_document_extractors()
}
\value{
Invisible NULL on success.
}
\description{
Clear all registered document extractors
}
