% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{clear_ocr_backends}
\alias{clear_ocr_backends}
\title{Clear all registered OCR backends}
\usage{
clear_ocr_backends()
}
\value{
Invisible NULL on success.
}
\description{
Clear all registered OCR backends
}
