% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{clear_post_processors}
\alias{clear_post_processors}
\title{Clear all registered post-processors}
\usage{
clear_post_processors()
}
\value{
Invisible NULL on success.
}
\description{
Clear all registered post-processors
}
