% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{detect_mime_type}
\alias{detect_mime_type}
\title{Detect MIME type from raw bytes}
\usage{
detect_mime_type(data)
}
\arguments{
\item{data}{Raw vector of bytes.}
}
\value{
Character string with detected MIME type.
}
\description{
Analyzes the content of a raw byte vector to determine the MIME type.
}
