% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{list_document_extractors}
\alias{list_document_extractors}
\title{List registered document extractors}
\usage{
list_document_extractors()
}
\value{
Character vector of registered extractor names.
}
\description{
List registered document extractors
}
