% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{list_post_processors}
\alias{list_post_processors}
\title{List registered post-processors}
\usage{
list_post_processors()
}
\value{
Character vector of registered post-processor names.
}
\description{
List registered post-processors
}
