% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{mime_type}
\alias{mime_type}
\title{Get the MIME type from a result}
\usage{
mime_type(x, ...)
}
\arguments{
\item{x}{A \code{kreuzberg_result} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Character string of the MIME type.
}
\description{
Get the MIME type from a result
}
