% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{print.kreuzberg_result}
\alias{print.kreuzberg_result}
\title{Print method for kreuzberg_result}
\usage{
\method{print}{kreuzberg_result}(x, ...)
}
\arguments{
\item{x}{A \code{kreuzberg_result} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for kreuzberg_result
}
