% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{validate_mime_type}
\alias{validate_mime_type}
\title{Validate a MIME type string}
\usage{
validate_mime_type(mime_type)
}
\arguments{
\item{mime_type}{Character string MIME type to validate.}
}
\value{
Logical indicating if the MIME type is valid.
}
\description{
Checks whether a given string is a valid and supported MIME type.
}
